refactor: re-structure the project
