+++ packages/rs-json-schema-compatibility/src/keyword/mod.rs
@@ -2,5 +2,5 @@ mod keyword_rule;
 mod rules;
 mod value;
 
-pub use keyword_rule::*;
+pub use keyword_rule::{KeywordRule, KeywordRuleExample, ReplaceCallback};
 pub use rules::KEYWORD_RULES;
